backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.0
pytest==7.4.3
jinja2==3.1.2
mysql-connector-python==8.2.0
pymysql==1.1.0

# GECToR 所需依赖
# torch>=1.10.0,<1.12.0  # 使用系统已安装的版本
transformers>=4.12.0
# allennlp>=2.9.0  # 由于版本冲突，暂时注释掉
nltk>=3.6.5
sentencepiece>=0.1.96
numpy>=1.20.0
ipython>=7.16.1